Commons: (to get the compilation working)          PARIZ.H  : increased for larger maps used for the warm snake          MAXTRA.H : reduced for the 2GB limit          MXLD.H   : reduced for the 2GB limit
For the management of maps
new data passed to KSMAP4 to decide if "NEWFIC". While KSMAP4 was only
using the file name,  to decide that we now include the coefficient
applied to the map. This forces modifications in toscac.f (where
coefficients are used) but also in all subroutines calling for KSMAP4,
in which we created dummy arguments(like constant coefficients) to be
passed to KSMAP4

  ksmap.f
  toscac.f
  emmac.f
  polmes.f
  map2d.f
  toscap.f

Added "IT" as argument when FITMM2 is called from integr.f