Various modifications to zgoubi/block.f.
For the benefit of future people tasked with updating the physical
constants, I added a lengthy comment concerning the computation
of nucleon magnetic moment anomalies from information listed in
the database of CODATA Recommended Values.

I added/modified comments about the data.

I did not change any of the data values, but added lines, currently
commented out, that contain new vlaues corrected on the basis of
the 2014 CODATA Recommended Values. My thinking is that we delete
the old values and uncomment the new values all at once, at some
mutually convenient time in the near future.