Remove man-page check on ndiff identity
Currently, ndiff is run from inside its build directory because of
problems with installing it in the docker image.  Therefore, there
is no man page.  To avoid confusion, it will be good to devise another
way of verifying that the correct ndiff is being used because APT
installs another package with the same name.