Comment unused code that GCC 9 correctly rejects
The top-level CMakeLists.txt invokes gfortran -fno-automatic,
causing GCC 9 to correctly reject lines 78-84 in
modules/particle_impl.f90 because the -fno-automatic flag gives
all variables the SAVE attribute, which is not allowed in PURE
procedures.  The aforementioned lines are not invoked in the
current tests, but could prove useful during upcoming performance
engineering efforts. Uncomment the aforementioned the
aforementioned lines to expose the compiler error.